+++ Changelog.md
@@ -2,6 +2,8 @@
 + Refactored `MediaAsset.Caption<T>` method to `MediaAsset.Parse<T>` for structured parsing.
 + Refactored `MediaAsset.Caption` method to `MediaAsset.Transcribe` for performing speech-to-text.
 + Refactored `MediaAsset.FromSpeechPrompt` method to `MediaAsset.FromGeneratedSpeech` for performing text-to-speech.
++ Removed `PixelBuffer.Region` method.
++ Removed `PixelBuffer.ToImage` method.
 
 ## 0.0.21
 + Fixed `WebException: The request was aborted: The request was canceled` when building for Android.